8219612: compiler.codecache.stress.Helper.TestCaseImpl can't be defined in different runtime package as its nest host Summary: Test rewritten to use top level classes rather then nested ones Reviewed-by: iignatyev, kvn, thartmann
