+++ test/hotspot/jtreg/compiler/codecache/stress/TestCaseImpl.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2014, 2019, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+
+package compiler.codecache.stress;
+
+import java.util.concurrent.Callable;
+
+public class TestCaseImpl implements Helper.TestCase {
+    private static final int RETURN_VALUE = 42;
+    private static final int RECURSION_DEPTH = 10;
+    private volatile int i;
+
+    @Override
+    public Callable<Integer> getCallable() {
+        return () -> {
+            i = 0;
+            return method();
+        };
+    }
+
+    @Override
+    public int method() {
+        ++i;
+        int result = RETURN_VALUE;
+        if (i < RECURSION_DEPTH) {
+            return result + method();
+        }
+        return result;
+    }
+
+    @Override
+    public int expectedValue() {
+        return RETURN_VALUE * RECURSION_DEPTH;
+    }
+}
